Add some :? to vars that may accidentally touch the host
A lot of this is redundant or not needed as there are not-empty checks
for the specific cases. However until we use set -u we should adopt this
style to avoid any mistakes from typos or refactoring.

Issue #977

In some cases ? is used as the command may be ran in the host.

+++ src/share/poudriere/distclean.sh
@@ -156,12 +156,11 @@ for PTNAME in ${PTNAMES}; do
 		echo "PACKAGE_BUILDING_FLAVORS=yes"
 	fi >> "${__MAKE_CONF}"
 	unset P_PORTS_FEATURES
-	fetch_global_port_vars
-
+	MASTERMNT= fetch_global_port_vars
 	MASTERMNT= MASTERMNTREL= load_moved
 	msg "Gathering all expected distfiles for ports tree '${PTNAME}'"
 
-	ports="$(listed_ports show_moved)" ||
+	ports="$(MASTERMNTREL= listed_ports show_moved)" ||
 	    err "$?" "Failed to find ports for ${PTNAME}"
 	for originspec in ${ports}; do
 		parallel_run \
@@ -184,7 +183,7 @@ msg "Gathering list of actual distfiles"
 # This is redundant but here for paranoia.
 [ -n "${DISTFILES_CACHE}" ] ||
     err 1 "DISTFILES_CACHE must be set (cf. poudriere.conf)"
-find -x ${DISTFILES_CACHE}/ -type f ! -name '.*' | \
+find -x ${DISTFILES_CACHE:?}/ -type f ! -name '.*' | \
     sort -o "${DISTFILES_LIST}.actual"
 
 comm -1 -3 ${DISTFILES_LIST}.expected ${DISTFILES_LIST}.actual \
@@ -200,5 +199,5 @@ if [ ${ret} -eq 2 ]; then
 	exit 0
 fi
 if [ "${DRY_RUN}" -eq 0 ]; then
-	find -x ${DISTFILES_CACHE}/ -type d -mindepth 1 -empty -delete
+	find -x ${DISTFILES_CACHE:?}/ -type d -mindepth 1 -empty -delete
 fi
